Make DType enum a static inner class of NDArray
The DType and Shape classes define the two major pieces of metadata
needed by NDArray. One was an inner class of NDArray and the other
was not. This change makes it consistent: now both are inner classes.
